requirements.txt
accelerate==1.2.1
bitsandbytes==0.45.0
datasets==3.2.0
diffusers==0.31.0
huggingface_hub==0.26.5
matplotlib==3.10.0
numpy==2.2.0
omegaconf==2.3.0
opencv_python==4.10.0.84
packaging==24.2
pandas==2.2.3
Pillow==11.0.0
PyYAML==6.0.2
torch_geometric==2.6.1
tqdm==4.67.1
transformers==4.47.0
wandb==0.19.1
xformers==0.0.28.post3
